Extract ImapPusher from ImapStore
